Refactoring based on the reviewer's suggestions
