+++ picky-asn1-x509/src/oids.rs
@@ -57,6 +57,12 @@ define_oid! {
     EMAIL_ADDRESS => email_address => "1.2.840.113549.1.9.1", // deprecated
     EXTENSION_REQ => extension_request => "1.2.840.113549.1.9.14",
 
+    // pkcs7
+    PKCS7 => pkcs7 => "1.2.840.113549.1.7.1",
+    SIGNED_DATA => signed_data => "1.2.840.113549.1.7.2",
+    CONTENT_TYPE => content_type => "1.2.840.113549.1.9.3",
+    MESSAGE_DIGEST => message_digest => "1.2.840.113549.1.9.4",
+
     // NIST
     DSA_WITH_SHA224 => dsa_with_sha224 => "2.16.840.1.101.3.4.3.1",
     DSA_WITH_SHA256 => dsa_with_sha256 => "2.16.840.1.101.3.4.3.2",
@@ -152,12 +158,6 @@ define_oid! {
     SHA3_512 => sha3_512 => "2.16.840.1.101.3.4.2.10",
     SHAKE128 => shake128 => "2.16.840.1.101.3.4.2.11",
     SHAKE256 => shake256 => "2.16.840.1.101.3.4.2.12",
-
-    // pkcs7
-    PKCS7 => pkcs7 => "1.2.840.113549.1.7.1",
-    SIGNED_DATA => signed_data => "1.2.840.113549.1.7.2",
-    CONTENT_TYPE => content_type => "1.2.840.113549.1.9.3",
-    MESSAGE_DIGEST => message_digest => "1.2.840.113549.1.9.4",
 
     // authenticode
     SPC_INDIRECT_DATA_OBJID => spc_indirect_data_objid => "1.3.6.1.4.1.311.2.1.4",
