+++ picky/src/hash.rs
@@ -3,6 +3,23 @@
 use digest::Digest;
 use picky_asn1_x509::SHAVariant;
 use serde::{Deserialize, Serialize};
+use std::convert::TryFrom;
+use std::error::Error;
+use std::fmt;
+
+/// unsupported algorithm
+#[derive(Debug)]
+pub struct UnsupportedHashAlgorithmError {
+    pub algorithm: String,
+}
+
+impl fmt::Display for UnsupportedHashAlgorithmError {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        write!(f, "unsupported algorithm:  {}", self.algorithm)
+    }
+}
+
+impl Error for UnsupportedHashAlgorithmError {}
 
 /// Supported hash algorithms
 #[derive(Deserialize, Serialize, Debug, Clone, Copy, PartialEq, Eq, Hash)]
@@ -30,28 +47,36 @@ impl From<HashAlgorithm> for rsa::Hash {
         }
     }
 }
-impl From<HashAlgorithm> for SHAVariant {
-    fn from(v: HashAlgorithm) -> SHAVariant {
+impl TryFrom<HashAlgorithm> for SHAVariant {
+    type Error = UnsupportedHashAlgorithmError;
+
+    fn try_from(v: HashAlgorithm) -> Result<SHAVariant, UnsupportedHashAlgorithmError> {
         match v {
-            HashAlgorithm::SHA2_256 => SHAVariant::SHA2_256,
-            HashAlgorithm::SHA2_384 => SHAVariant::SHA2_384,
-            HashAlgorithm::SHA2_512 => SHAVariant::SHA2_512,
-            HashAlgorithm::SHA3_384 => SHAVariant::SHA3_384,
-            HashAlgorithm::SHA3_512 => SHAVariant::SHA3_512,
-            _ => SHAVariant::SHA2_256,
+            HashAlgorithm::SHA2_256 => Ok(SHAVariant::SHA2_256),
+            HashAlgorithm::SHA2_384 => Ok(SHAVariant::SHA2_384),
+            HashAlgorithm::SHA2_512 => Ok(SHAVariant::SHA2_512),
+            HashAlgorithm::SHA3_384 => Ok(SHAVariant::SHA3_384),
+            HashAlgorithm::SHA3_512 => Ok(SHAVariant::SHA3_512),
+            _ => Err(UnsupportedHashAlgorithmError {
+                algorithm: format!("{:?}", v),
+            }),
         }
     }
 }
 
-impl From<SHAVariant> for HashAlgorithm {
-    fn from(v: SHAVariant) -> HashAlgorithm {
+impl TryFrom<SHAVariant> for HashAlgorithm {
+    type Error = UnsupportedHashAlgorithmError;
+
+    fn try_from(v: SHAVariant) -> Result<HashAlgorithm, UnsupportedHashAlgorithmError> {
         match v {
-            SHAVariant::SHA2_256 => HashAlgorithm::SHA2_256,
-            SHAVariant::SHA2_384 => HashAlgorithm::SHA2_384,
-            SHAVariant::SHA2_512 => HashAlgorithm::SHA2_512,
-            SHAVariant::SHA3_384 => HashAlgorithm::SHA3_384,
-            SHAVariant::SHA3_512 => HashAlgorithm::SHA3_512,
-            _ => HashAlgorithm::SHA2_256,
+            SHAVariant::SHA2_256 => Ok(HashAlgorithm::SHA2_256),
+            SHAVariant::SHA2_384 => Ok(HashAlgorithm::SHA2_384),
+            SHAVariant::SHA2_512 => Ok(HashAlgorithm::SHA2_512),
+            SHAVariant::SHA3_384 => Ok(HashAlgorithm::SHA3_384),
+            SHAVariant::SHA3_512 => Ok(HashAlgorithm::SHA3_512),
+            _ => Err(UnsupportedHashAlgorithmError {
+                algorithm: format!("{:?}", v),
+            }),
         }
     }
 }
