+++ picky/src/lib.rs
@@ -27,6 +27,7 @@ mod test_files {
     pub const RSA_2048_PK_7: &str = include_str!("../../test_assets/private_keys/rsa-2048-pk_7.key");
     pub const RSA_4096_PK_3: &str = include_str!("../../test_assets/private_keys/rsa-4096-pk_3.key");
 
+    #[cfg(feature = "pkcs7")]
     pub const PKCS7: &str = include_str!("../../test_assets/pkcs7.p7b");
 
     cfg_if::cfg_if! { if #[cfg(feature = "x509")] {
