+++ picky/src/x509/utils.rs
@@ -3,9 +3,7 @@ use serde::{Deserialize, Serialize};
 use super::certificate::CertError;
 use crate::pem::{parse_pem, Pem};
 
-pub type UtilsResult<T> = Result<T, CertError>;
-
-pub(super) fn from_der<'a, T, V>(data: &'a V, element: &'static str) -> UtilsResult<T>
+pub(super) fn from_der<'a, T, V>(data: &'a V, element: &'static str) -> Result<T, CertError>
 where
     T: Deserialize<'a>,
     V: ?Sized + AsRef<[u8]>,
@@ -13,33 +11,36 @@ where
     picky_asn1_der::from_bytes(data.as_ref()).map_err(|e| CertError::Asn1Deserialization { source: e, element })
 }
 
-#[allow(unreachable_patterns)] // Rust compiler thinks that this is unreachable pattern, but it's reachable pattern.
-                               // it's unreachable only if the function called in from_pem_str().
 pub(super) fn from_pem<'a, T: Deserialize<'a>>(
     pem: &'a Pem,
-    _pem_label: &'a str,
+    expected_pem_label: &'a str,
     element: &'static str,
-) -> UtilsResult<T> {
-    match pem.label() {
-        _pem_label => from_der(pem.data(), element),
-        _ => Err(CertError::InvalidPemLabel {
+) -> Result<T, CertError> {
+    if pem.label() == expected_pem_label {
+        from_der(pem.data(), element)
+    } else {
+        Err(CertError::InvalidPemLabel {
             label: pem.label().to_owned(),
-        }),
+        })
     }
 }
 
-pub(super) fn from_pem_str<T>(pem_str: &str, pem_label: &str, element: &'static str) -> UtilsResult<T>
+pub(super) fn from_pem_str<T>(pem_str: &str, expected_pem_label: &str, element: &'static str) -> Result<T, CertError>
 where
     for<'a> T: Deserialize<'a>,
 {
     let pem = parse_pem(pem_str)?;
-    from_pem(&pem, pem_label, element)
+    from_pem(&pem, expected_pem_label, element)
 }
 
-pub(super) fn to_der<T: Serialize>(val: &T, element: &'static str) -> UtilsResult<Vec<u8>> {
+pub(super) fn to_der<T: Serialize>(val: &T, element: &'static str) -> Result<Vec<u8>, CertError> {
     picky_asn1_der::to_vec(val).map_err(|e| CertError::Asn1Serialization { source: e, element })
 }
 
-pub(super) fn to_pem<T: Serialize>(val: &T, pem_label: &str, element: &'static str) -> UtilsResult<Pem<'static>> {
-    Ok(Pem::new(pem_label, to_der(val, element)?))
+pub(super) fn to_pem<T: Serialize>(
+    val: &T,
+    expected_pem_label: &str,
+    element: &'static str,
+) -> Result<Pem<'static>, CertError> {
+    Ok(Pem::new(expected_pem_label, to_der(val, element)?))
 }
